Fix on order domain

+++ services/order-service/src/main/java/com/denyaar/orderservice/product/PurchaseRequest.java
@@ -15,7 +15,7 @@ public record PurchaseRequest(
         Integer productId,
         @NotNull(message = "Quantity is required")
         @Positive(message = "Quantity must be greater than zero")
-        Integer quantity
+        double quantity
 
 ) {
 }
